requirements.txt
python-consul>=0.7.0
pypureomapi>=0.6
flask>=0.12.2
